+++ Source/GA6thFinal_Framework/GameEngine/GameEngine.vcxproj
@@ -285,6 +285,11 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="Source\Engine\GraphicsCore\ComputeShader.h" />
+    <ClInclude Include="Source\Engine\GraphicsCore\GeometryShader.h" />
+    <ClInclude Include="Source\Engine\GraphicsCore\MeshShader.h" />
+    <ClInclude Include="Source\Engine\GraphicsCore\PixelShader.h" />
+    <ClInclude Include="Source\Engine\GraphicsCore\Shader.h" />
     <ClInclude Include="Source\Editor\Menu\Project\EditorBuildSettingMenu.h" />
     <ClInclude Include="Source\Engine\GraphicsCore\DeferredPBRLitPass.h" />
     <ClInclude Include="Source\Engine\GraphicsCore\GBufferPass.h" />
@@ -358,7 +363,7 @@
     <ClInclude Include="Source\Engine\GraphicsCore\RenderTarget.h" />
     <ClInclude Include="Source\Engine\GraphicsCore\Resource.h" />
     <ClInclude Include="Source\Engine\GraphicsCore\ResourceManager.h" />
-    <ClInclude Include="Source\Engine\GraphicsCore\Shader.h" />
+    <ClInclude Include="Source\Engine\GraphicsCore\ShaderBuilder.h" />
     <ClInclude Include="Source\Engine\GraphicsCore\Skeleton.h" />
     <ClInclude Include="Source\Engine\GraphicsCore\Sphere.h" />
     <ClInclude Include="Source\Engine\GraphicsCore\Structs.h" />
@@ -409,8 +414,14 @@
     <ClInclude Include="Source\pch.h" />
     <ClInclude Include="Source\Engine\Utility\DumpUtility.h" />
     <ClInclude Include="Source\UmScripts.h" />
+    <ClInclude Include="Source\Engine\GraphicsCore\VertexShader.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="Source\Engine\GraphicsCore\ComputeShader.cpp" />
+    <ClCompile Include="Source\Engine\GraphicsCore\GeometryShader.cpp" />
+    <ClCompile Include="Source\Engine\GraphicsCore\MeshShader.cpp" />
+    <ClCompile Include="Source\Engine\GraphicsCore\PixelShader.cpp" />
+    <ClCompile Include="Source\Engine\GraphicsCore\Shader.cpp" />
     <ClCompile Include="Source\Editor\Menu\Project\EditorBuildSettingMenu.cpp" />
     <ClCompile Include="Source\Engine\GraphicsCore\DeferredPBRLitPass.cpp" />
     <ClCompile Include="Source\Engine\GraphicsCore\GBufferPass.cpp" />
@@ -514,9 +525,8 @@
     <ClCompile Include="Source\Engine\GraphicsCore\Quad.cpp" />
     <ClCompile Include="Source\Engine\GraphicsCore\Renderer.cpp" />
     <ClCompile Include="Source\Engine\GraphicsCore\RenderTarget.cpp" />
-    <ClCompile Include="Source\Engine\GraphicsCore\Resource.cpp" />
     <ClCompile Include="Source\Engine\GraphicsCore\ResourceManager.cpp" />
-    <ClCompile Include="Source\Engine\GraphicsCore\Shader.cpp" />
+    <ClCompile Include="Source\Engine\GraphicsCore\ShaderBuilder.cpp" />
     <ClCompile Include="Source\Engine\GraphicsCore\Skeleton.cpp" />
     <ClCompile Include="Source\Engine\GraphicsCore\Sphere.cpp" />
     <ClCompile Include="Source\Engine\GraphicsCore\StructuredBuffer.cpp" />
@@ -620,6 +630,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='ReleaseEditor|x64'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="Source\Engine\Utility\DumpUtility.cpp" />
+    <ClCompile Include="Source\Engine\GraphicsCore\VertexShader.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="cpp.hint" />
