+++ Source/GA6thFinal_Framework/GameEngine/GameEngine.vcxproj.filters
@@ -94,6 +94,9 @@
     <Filter Include="GraphicsCore\Material">
       <UniqueIdentifier>{cc2e0fd2-4be0-475a-989a-1b89bb9ef427}</UniqueIdentifier>
     </Filter>
+    <Filter Include="GraphicsCore\Resource\ShaderBlob">
+      <UniqueIdentifier>{269d6efd-b9fd-480e-903e-d56777bc8493}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Source\pch.h">
@@ -381,7 +384,7 @@
     <ClInclude Include="Source\Engine\GraphicsCore\Renderer.h">
       <Filter>GraphicsCore\Renderer</Filter>
     </ClInclude>
-    <ClInclude Include="Source\Engine\GraphicsCore\Shader.h">
+    <ClInclude Include="Source\Engine\GraphicsCore\ShaderBuilder.h">
       <Filter>GraphicsCore\Shader</Filter>
     </ClInclude>
     <ClInclude Include="Source\Engine\AppModule\ImGuiDX12Module.h">
@@ -450,12 +453,6 @@
     <ClInclude Include="Source\Editor\Menu\Project\EditorBuildSettingMenu.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="Source\Engine\GraphicsCore\PBRPass.h">
-      <Filter>헤더 파일</Filter>
-    </ClInclude>
-    <ClInclude Include="Source\Engine\GraphicsCore\PBRTechnique.h">
-      <Filter>헤더 파일</Filter>
-    </ClInclude>
     <ClInclude Include="Source\Engine\GraphicsCore\Animator.h">
       <Filter>GraphicsCore\Animator</Filter>
     </ClInclude>
@@ -472,6 +469,24 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="Source\Engine\GraphicsCore\DeferredPBRLitPass.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\Engine\GraphicsCore\Shader.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\Engine\GraphicsCore\VertexShader.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\Engine\GraphicsCore\ComputeShader.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\Engine\GraphicsCore\GeometryShader.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\Engine\GraphicsCore\MeshShader.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\Engine\GraphicsCore\PixelShader.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
@@ -719,7 +734,7 @@
     <ClCompile Include="Source\Engine\GraphicsCore\Renderer.cpp">
       <Filter>GraphicsCore\Renderer</Filter>
     </ClCompile>
-    <ClCompile Include="Source\Engine\GraphicsCore\Shader.cpp">
+    <ClCompile Include="Source\Engine\GraphicsCore\ShaderBuilder.cpp">
       <Filter>GraphicsCore\Shader</Filter>
     </ClCompile>
     <ClCompile Include="Source\Engine\AppModule\ImGuiDX12Module.cpp">
@@ -742,9 +757,6 @@
     </ClCompile>
     <ClCompile Include="Source\Engine\GraphicsCore\GraphicsModule.cpp">
       <Filter>GraphicsCore\Module</Filter>
-    </ClCompile>
-    <ClCompile Include="Source\Engine\GraphicsCore\Resource.cpp">
-      <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="Source\Engine\GraphicsCore\RenderScene.cpp">
       <Filter>소스 파일</Filter>
@@ -788,9 +800,6 @@
     <ClCompile Include="Source\Editor\Menu\Project\EditorBuildSettingMenu.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="Source\Engine\GraphicsCore\PBRPass.cpp">
-      <Filter>소스 파일</Filter>
-    </ClCompile>
     <ClCompile Include="Source\Engine\GraphicsCore\Animator.cpp">
       <Filter>GraphicsCore\Animator</Filter>
     </ClCompile>
@@ -807,6 +816,24 @@
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="Source\Engine\GraphicsCore\DeferredPBRLitPass.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\Engine\GraphicsCore\Shader.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\Engine\GraphicsCore\VertexShader.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\Engine\GraphicsCore\ComputeShader.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\Engine\GraphicsCore\GeometryShader.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\Engine\GraphicsCore\MeshShader.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\Engine\GraphicsCore\PixelShader.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
   </ItemGroup>
