+++ Source/GA6thFinal_Framework/GameEngine/Source/Engine/GraphicsCore/GeometryShader.cpp
@@ -1,0 +1,10 @@
+﻿#include "pch.h"
+#include "GeometryShader.h"
+
+HRESULT GeometryShader::LoadResource(const std::filesystem::path& filePath)
+{
+    HRESULT hr = CompileShader(filePath.c_str(), "gs_main", "gs_5_1");
+    FAILED_CHECK_BREAK(hr);
+
+    return hr;
+}
