--- Source/GA6thFinal_Framework/GameEngine/Source/Engine/GraphicsCore/Resource.cpp
@@ -1,7 +1,0 @@
-﻿#include "pch.h"
-
-HRESULT Resource::LoadResource(const std::filesystem::path& filePath)
-{
-    ASSERT(false, L"상속 받아서 사용해주세요");
-    return E_FAIL;
-}
